Fix install script - remove reference to non-existent wifi_monitor.py
The wifi_monitor.py file was removed during cleanup since we create
monitor.py directly in the install script. This fixes the error:
'cp: cannot stat wifi_monitor.py: No such file or directory'

The monitor.py is created as a here-document in the install script
itself, so no file needs to be copied.